optimization-methods/01-grad-descent/requirements.txt
numpy
matplotlib
jupyter
Ipython
opencv-python
Pillow
imageio